+++ anki/public/index.css
@@ -20,6 +20,10 @@ body {
     height: 100%;
 }
 
+button {
+    cursor: pointer;
+}
+
 .shadow-out-bottom {
     box-shadow: 0px 4px 4px #000000;
 }
@@ -30,4 +34,12 @@ body {
 
 .shadow-in-top-light {
     box-shadow: inset 0px 4px 4px rgba(0, 0, 0, 0.25);
+}
+
+.button-switch {
+    width: 90%;
+}
+
+.button-switch:hover:enabled {
+    width: 100%;
 }
