+++ anki/src/components/ButtonSwitch.tsx
@@ -6,7 +6,8 @@ interface Props {
     text: string,
     width: number,
     height: number,
-    onClick: () => void | undefined,
+    fontSize?: number,
+    onClick?: () => void,
     onMouseDown?: () => void,
     onMouseUp?: () => void,
 }
@@ -14,20 +15,26 @@ interface Props {
 const ButtonSwitch = (props: Props) => {
     const [theme, ] = useContext(ThemeContext)
     return(
-        <button disabled={props.is_on} className={props.is_on ? 'shadow-in-top' : 'shadow-out-bottom'} style={{
+        <div style={{
             width: props.width,
             height: props.height,
-            backgroundColor: props.is_on ? theme.button_switch_on : theme.button_switch_off,
-            color: props.is_on ? '#FFFFFF' : '#696969',
-            fontSize: '20px',
-            border: 'none',
-            borderRadius: props.height / 2,
-            transition: 'background-color 100ms',
-        }} onClick={props.onClick}
-           onMouseDown={props.onMouseDown}
-           onMouseUp={props.onMouseUp}>
-            {props.text}
-        </button>
+            display: 'flex',
+            justifyContent: 'center',
+        }}>
+            <button disabled={props.is_on} className={'button-switch '.concat(props.is_on ? 'shadow-in-top' : 'shadow-out-bottom')} style={{
+                height: props.height,
+                backgroundColor: props.is_on ? theme.button_switch_on : theme.button_switch_off,
+                color: props.is_on ? '#FFFFFF' : '#696969',
+                fontSize: props.fontSize ? props.fontSize : '20px',
+                border: 'none',
+                borderRadius: props.height / 2,
+                transition: 'background-color 100ms, width 100ms',
+            }} onClick={props.onClick}
+            onMouseDown={props.onMouseDown}
+            onMouseUp={props.onMouseUp}>
+                {props.text}
+            </button>
+        </div>
     )
 };
