+++ anki/src/components/Header.tsx
@@ -22,5 +22,4 @@ const Header = () => {
     )
 };
 
-
 export default Header
